+++ src/components/TodoInfo/TodoInfo.tsx
@@ -1,6 +1,5 @@
-import React from 'react';
-import { Todo } from '../../types/TODO';
 import classNames from 'classnames';
+import { Todo } from '../../types/Todo';
 import { UserInfo } from '../UserInfo';
 
 type Props = {
@@ -16,7 +15,6 @@ export const TodoInfo: React.FC<Props> = ({ todo }) => {
       })}
     >
       <h2 className="TodoInfo__title">{todo.title}</h2>
-
       {todo.user && <UserInfo user={todo.user} />}
     </article>
   );
