+++ src/components/TodoList/TodoList.tsx
@@ -1,5 +1,5 @@
+import { Todo } from '../../types/Todo';
 import React from 'react';
-import { Todo } from '../../types/TODO';
 import { TodoInfo } from '../TodoInfo';
 
 type Props = {
